requirements.txt
git+https://github.com/tall-josh/yamgal.git@wip
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.0
click==7.1.2
cssselect==1.1.0
cssselect2==0.3.0
defusedxml==0.6.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.0
MarkupSafe==1.1.1
Pillow==7.1.2
pycparser==2.20
pygal==2.4.0
PyYAML==5.3.1
tinycss==0.4
tinycss2==1.0.2
urllib3==1.25.9
webencodings==0.5.1
Werkzeug==1.0.1
gunicorn==20.0.4
